changed to double

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -8,7 +8,7 @@ public class MyMathCalc {
     public int sub(int a, int b){
         return a - b;
     }
-    public int div(int a, int b){
+    public double div(double a, double b){
         return a / b;
     }
 }
